refactor: Merge enums
